# backend_golang/docs/swagger.yaml
basePath: /api/v1
definitions:
  ent.Appointment:
    properties:
      DoctorId:
        description: DoctorId holds the value of the "DoctorId" field.
        type: integer
      UserId:
        description: UserId holds the value of the "UserId" field.
        type: integer
      detail:
        description: Detail holds the value of the "detail" field.
        type: string
      edges:
        $ref: '#/definitions/ent.AppointmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AppointmentQuery when eager-loading is set.
      endTime:
        description: EndTime holds the value of the "endTime" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      reasonForAppointment:
        description: ReasonForAppointment holds the value of the "reasonForAppointment"
          field.
        type: string
      startTime:
        description: StartTime holds the value of the "startTime" field.
        type: string
      status:
        description: Status holds the value of the "status" field.
        type: string
    type: object
...
  ent.Certification:
    properties:
      Issuer:
        description: Issuer holds the value of the "Issuer" field.
        type: string
      code:
        description: Code holds the value of the "code" field.
        type: string
      dateOfExp:
        description: DateOfExp holds the value of the "dateOfExp" field.
        type: string
      dateOfIssuing:
        description: DateOfIssuing holds the value of the "dateOfIssuing" field.
        type: string
      diloma:
        description: Diloma holds the value of the "diloma" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CertificationEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CertificationQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CertificationEdges:
    properties:
      doctor_owner:
        $ref: '#/definitions/ent.User'
        description: DoctorOwner holds the value of the doctor_owner edge.
    type: object
  ent.Chat:
    properties:
      Chat_room_name:
        description: ChatRoomName holds the value of the "Chat_room_name" field.
        type: string
      IsLockChat:
        description: IsLockChat holds the value of the "IsLockChat" field.
        type: boolean
      edges:
        $ref: '#/definitions/ent.ChatEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ChatQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ChatEdges:
    properties:
      chat_appointment:
        description: ChatAppointment holds the value of the chat_appointment edge.
        items:
          $ref: '#/definitions/ent.Appointment'
        type: array
      chat_message:
        description: ChatMessage holds the value of the chat_message edge.
        items:
          $ref: '#/definitions/ent.Message'
        type: array
      chat_user:
        description: ChatUser holds the value of the chat_user edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.Department:
    properties:
      edges:
        $ref: '#/definitions/ent.DepartmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DepartmentQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      image:
        description: Image holds the value of the "image" field.
        type: string
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
...
  ent.Disease:
    properties:
      edges:
        $ref: '#/definitions/ent.DiseaseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DiseaseQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      symtoms:
        description: Symtoms holds the value of the "symtoms" field.
        type: string
    type: object
...
  ent.Hospital:
    properties:
      edges:
        $ref: '#/definitions/ent.HospitalEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the HospitalQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
...
  ent.Message:
    properties:
      edges:
        $ref: '#/definitions/ent.MessageEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MessageQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      message_text:
        description: MessageText holds the value of the "message_text" field.
        type: string
      sent_dateTime:
        description: SentDateTime holds the value of the "sent_dateTime" field.
        type: string
    type: object
  ent.MessageEdges:
    properties:
      What_messages_are_in_this_chat:
        $ref: '#/definitions/ent.Chat'
        description: WhatMessagesAreInThisChat holds the value of the What_messages_are_in_this_chat
          edge.
      Who_send_messages:
        $ref: '#/definitions/ent.User'
        description: WhoSendMessages holds the value of the Who_send_messages edge.
    type: object
  ent.Notification:
    properties:
      AppointmentId:
        description: AppointmentId holds the value of the "AppointmentId" field.
        type: integer
      CreatedDate:
        description: CreatedDate holds the value of the "CreatedDate" field.
        type: string
      Message:
        description: Message holds the value of the "Message" field.
        type: string
      RecipientId:
        description: RecipientId holds the value of the "RecipientId" field.
        type: integer
      SenderId:
        description: SenderId holds the value of the "SenderId" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.NotificationEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the NotificationQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.NotificationEdges:
    properties:
      user_notification:
        description: UserNotification holds the value of the user_notification edge.
        items:
          $ref: '#/definitions/ent.User'
        type: array
    type: object
  ent.PInfo:
    properties:
      address:
        description: Address holds the value of the "address" field.
        type: string
      bloodGroup:
        description: BloodGroup holds the value of the "bloodGroup" field.
        type: string
      brithDate:
        description: BrithDate holds the value of the "brithDate" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PInfoEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PInfoQuery when eager-loading is set.
      firstName:
        description: FirstName holds the value of the "firstName" field.
        type: string
      gender:
        description: Gender holds the value of the "gender" field.
        type: integer
      id:
        description: ID of the ent.
        type: integer
      idCardNumber:
        description: IdCardNumber holds the value of the "idCardNumber" field.
        type: string
      lastName:
        description: LastName holds the value of the "lastName" field.
        type: string
    type: object
...
  ent.Schedule:
    properties:
      activity:
        description: Activity holds the value of the "activity" field.
        type: string
      detail:
        description: Detail holds the value of the "detail" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ScheduleEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ScheduleQuery when eager-loading is set.
      endTime:
        description: EndTime holds the value of the "endTime" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      startTime:
        description: StartTime holds the value of the "startTime" field.
        type: string
      status:
        description: Status holds the value of the "status" field.
        type: string
    type: object
...
  ent.Treatment:
    properties:
      dateTime:
        description: DateTime holds the value of the "dateTime" field.
        type: string
      edges:
        $ref: '#/definitions/ent.TreatmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TreatmentQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      takeTime:
        description: TakeTime holds the value of the "takeTime" field.
        type: number
      treatmentRecord:
        description: TreatmentRecord holds the value of the "treatmentRecord" field.
        type: string
    type: object
...
  ent.User:
    properties:
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      password:
        description: Password holds the value of the "password" field.
        type: string
      username:
        description: Username holds the value of the "username" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      doctor_has_certification:
        description: DoctorHasCertification holds the value of the doctor_has_certification
          edge.
        items:
          $ref: '#/definitions/ent.Certification'
        type: array
      doctor_record_treatment:
        description: DoctorRecordTreatment holds the value of the doctor_record_treatment
          edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
      doctor_schedule:
        description: DoctorSchedule holds the value of the doctor_schedule edge.
        items:
          $ref: '#/definitions/ent.Schedule'
        type: array
      from_hospital:
        $ref: '#/definitions/ent.Hospital'
        description: FromHospital holds the value of the from_hospital edge.
      has_department:
        $ref: '#/definitions/ent.Department'
        description: HasDepartment holds the value of the has_department edge.
      user_has_PInfo:
        description: UserHasPInfo holds the value of the user_has_PInfo edge.
        items:
          $ref: '#/definitions/ent.PInfo'
        type: array
      user_have_disease:
        description: UserHaveDisease holds the value of the user_have_disease edge.
        items:
          $ref: '#/definitions/ent.Disease'
        type: array
      user_have_role:
        description: UserHaveRole holds the value of the user_have_role edge.
        items:
          $ref: '#/definitions/ent.Role'
        type: array
      user_have_telecoms:
        description: UserHaveTelecoms holds the value of the user_have_telecoms edge.
        items:
          $ref: '#/definitions/ent.Telecom'
        type: array
      user_have_token:
        description: UserHaveToken holds the value of the user_have_token edge.
        items:
          $ref: '#/definitions/ent.Token'
        type: array
      user_have_treatment:
        description: UserHaveTreatment holds the value of the user_have_treatment
          edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
      user_send_message:
        description: UserSendMessage holds the value of the user_send_message edge.
        items:
          $ref: '#/definitions/ent.Message'
        type: array
      user_send_notification:
        description: UserSendNotification holds the value of the user_send_notification
          edge.
        items:
          $ref: '#/definitions/ent.Notification'
        type: array
      who_is_in_this_chat:
        description: WhoIsInThisChat holds the value of the who_is_in_this_chat edge.
        items:
          $ref: '#/definitions/ent.Chat'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT FINAL PROJECT Example API
  version: "1.0"
paths:
...
  /messages:
    get:
      description: list message entities
      operationId: list-message
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List message entities
    post:
      consumes:
      - application/json
      description: Create message
      operationId: create-message
      parameters:
      - description: Message entity
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/ent.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create message
  /messages/{id}:
    delete:
      description: get message by ID
      operationId: delete-message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a message entity by ID
    get:
      description: get message by ID
      operationId: get-message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a message entity by ID
    put:
      consumes:
      - application/json
      description: update message by ID
      operationId: update-message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message entity
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/ent.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a message entity by ID
  /notifications:
    get:
      description: list notification entities
      operationId: list-notification
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Notification'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List notification entities
    post:
      consumes:
      - application/json
      description: Create notification
      operationId: create-notification
      parameters:
      - description: Notification entity
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/ent.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create notification
  /notifications/{id}:
    delete:
      description: get notification by ID
      operationId: delete-notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a notification entity by ID
    get:
      description: get notification by ID
      operationId: get-notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a notification entity by ID
    put:
      consumes:
      - application/json
      description: update notification by ID
      operationId: update-notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      - description: Notification entity
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/ent.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Notification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a notification entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"